+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Command usage always displays mandatory parameters before optional ones
   - Error output (i.e. "missing Commandfile") is now printed using
     the configured ui class (allows colored output)
+  - Invalid parameters are displayed in the error message
   - Missing parameters are displayed in the error message
   - Parameter wrapping is only done if a value is passed
     or a default (at least empty string) is configured